\name{exampledata}
\alias{exampledata}
\docType{data}
\title{
Hypothetical data for example
}
\description{
Hypothetical data for example
}
\usage{data("exampledata")}
\format{
\describe{
    \item{yob}{the primary outcome}
    \item{sob}{the surrogate marker}
    \item{aob}{the treatment group indicator where 1 is treatment and 0 is control}
    }
}
\examples{
data(exampledata)
}
